refactor: adjust contact service into its own package
